+++ src/com/genir/aitweaks/core/shipai/autofire/Tag.kt
@@ -5,14 +5,15 @@ import com.genir.starfarer.loading.LoadingUtils
 import org.json.JSONArray
 
 enum class Tag {
-    FINISHER_BEAM,      // Weapon is affected by the Finisher Beam Protocol hullmod.
-    APPROACH_CLOSER,    // Ships with Custom AI approach to 80% of the weapon maximum range.
-    ANTI_SHIELD,        // Weapon will attack enemy ships only when their shields are raised
-    TRIGGER_HAPPY,      // Weapon will use a specialized AI, which ignores some of the hold-fire rules.
-    NO_MODDED_AI,       // Weapon is not subject to AI Tweaks modded behavior.
-    ANTI_FIGHTER,       // Same as vanilla ANTO_FTR AI hint.
-    NO_STAGGERED_FIRE,  // Weapon will not use staggered firing mode.
-    USE_LESS_VS_SHIELDS // Same as vanilla USE_LESS_VS_SHIELDS; weapon will attack enemy ships only when their shields are down.
+    FINISHER_BEAM,       // Weapon is affected by the Finisher Beam Protocol hullmod.
+    APPROACH_CLOSER,     // Ships with Custom AI approach to 80% of the weapon maximum range.
+    ANTI_SHIELD,         // Weapon will attack enemy ships only when their shields are raised
+    TRIGGER_HAPPY,       // Weapon will use a specialized AI, which ignores some of the hold-fire rules.
+    NO_MODDED_AI,        // Weapon is not subject to AI Tweaks modded behavior.
+    ANTI_FIGHTER,        // Same as vanilla ANTO_FTR AI hint.
+    NO_STAGGERED_FIRE,   // Weapon will not use staggered firing mode.
+    USE_LESS_VS_SHIELDS, // Same as vanilla USE_LESS_VS_SHIELDS; weapon will attack enemy ships only when their shields are down.
+    PRIMARY_WEAPON       // Will cause the AI to treat a PD weapon as a primary ship weapon and approach enemies within that weapon’s range.
 }
 
 private val weaponTags: MutableMap<String, Set<Tag>> = mutableMapOf()